fix(CD): Fix top_stock_data_extractor_deploy.yml
- Changed branch to master
- Added the top_stock_data_extractor_deploy.yml path to trigger paths.
- Removed job with the name "Change directory to top_stock_data_extractor"
since this wasn't actually changing to that directory when we need it to
build the docker image. The directory is now being changed in the
build, tag, and push to Amazon ECR job.
- Fixed typo of lambda function in comments.
- Added --ouput type=docker which prevents the following error
WARNING: No output specified with docker-container driver. Build result
will only remain in the build cache. To push result image into registry
use --push or to load image into docker use --load. An image does not
exist locally with the tag: ***.dkr.ecr.***.amazonaws.com/***